Refactor trace writer cleanup helper
